fix large number printing and makefile using deleted interpreter src

+++ core/src/eval/value_t.h
@@ -3,7 +3,7 @@ typedef enum ValueType { VAL_NUMBER, VAL_BOOL, VAL_STRING, VAL_NULL } ValueType;
 typedef struct Value {
 	ValueType type;
 	union {
-		double number;
+		long double number;
 		int boolean;
 		char* string;
 	};
